+++ packages/connectors/connector-tencent-sms/src/constant.ts
@@ -4,7 +4,6 @@ import { ConnectorConfigFormItemType } from '@logto/connector-kit';
 export const defaultMetadata: ConnectorMetadata = {
   id: 'tencent-short-message-service',
   target: 'tencent-sms',
-  platform: null,
   name: {
     en: 'Tencent Short Message Service',
     'zh-CN': '腾讯云短信服务',
@@ -33,6 +32,7 @@ export const defaultMetadata: ConnectorMetadata = {
       label: 'Access Key Secret',
       type: ConnectorConfigFormItemType.Text,
       required: true,
+      isConfidential: true,
       placeholder: '<access-key-secret>',
     },
     {