+++ packages/connectors/connector-twilio-sms/src/constant.ts
@@ -6,7 +6,6 @@ export const endpoint = 'https://api.twilio.com/2010-04-01/Accounts/{{accountSID
 export const defaultMetadata: ConnectorMetadata = {
   id: 'twilio-short-message-service',
   target: 'twilio-sms',
-  platform: null,
   name: {
     en: 'Twilio SMS Service',
     'zh-CN': 'Twilio 短信服务',